requirements.txt
Flask
faiss-cpu
sentence-transformers
openai
googletrans==4.0.0-rc1
pandas
numpy
scikit-learn
transformers
torch
tqdm
requests
beautifulsoup4
langchain
pydantic
langchain-openai
langchain-community
langchain-text-splitters
langchain-embeddings
langchain-vectorstores
langchain-retrievers
langchain-chains
langchain-prompts
langchain-document-loaders
langchain-utilities